Add build flavors to split Amazon and Google SDK
