+++ library/build.gradle
@@ -23,6 +23,14 @@ android {
         versionName rootProject.ext.general.versionName
     }
 
+    flavorDimensions "appstore"
+    productFlavors {
+        google {
+        }
+        amazon {
+        }
+    }
+
     lintOptions {
         warningsAsErrors true
         // ok disables
@@ -32,12 +40,13 @@ android {
 }
 
 dependencies {
-    implementation "androidx.core:core-ktx:1.2.0"
+    implementation "androidx.core:core-ktx:1.3.1"
+    implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
-    api files('libs/in-app-purchasing-2.0.76.jar')
-    api 'androidx.appcompat:appcompat:1.1.0'
-    api 'com.android.billingclient:billing:2.2.0'
+    amazonImplementation files('libs/in-app-purchasing-2.0.76.jar')
+
+    googleImplementation 'com.android.billingclient:billing:3.0.0'
 
     testImplementation 'junit:junit:4.13'
     testImplementation 'org.mockito:mockito-core:2.24.0'
