+++ library/src/main/java/com/eggheadgames/inapppayments/IAPManager.kt
@@ -4,40 +4,25 @@ import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Context
 import com.billing.BillingService
+import com.billing.IBillingService
 import com.billing.PurchaseServiceListener
 import com.billing.SubscriptionServiceListener
-import com.billing.amazon.AmazonBillingService
-import com.billing.google.GoogleBillingService2
-import java.util.*
 
 //Public front-end for IAP functionality.
 object IAPManager {
-    const val BUILD_TARGET_GOOGLE = 0
-    const val BUILD_TARGET_AMAZON = 1
 
     @SuppressLint("StaticFieldLeak")
-    private var mBillingService: BillingService? = null
+    private var mBillingService: IBillingService? = null
 
     /**
      * @param context          - application context
-     * @param buildTarget      - IAPManager.BUILD_TARGET_GOOGLE or IAPManager.BUILD_TARGET_AMAZON
      * @param iapKeys          - list of sku for purchases
      * @param subscriptionKeys - list of sku for subscriptions
      */
     @JvmStatic
-    fun build(context: Context, buildTarget: Int, iapKeys: List<String>, subscriptionKeys: List<String> = emptyList()) {
+    fun build(context: Context, iapKeys: List<String>, subscriptionKeys: List<String> = emptyList()) {
         val contextLocal = context.applicationContext ?: context
-
-        //Build-specific initializations
-        if (buildTarget == BUILD_TARGET_GOOGLE) {
-            mBillingService = GoogleBillingService2(contextLocal, iapKeys, subscriptionKeys)
-
-        } else if (buildTarget == BUILD_TARGET_AMAZON) {
-            val keys: MutableList<String> = ArrayList()
-            keys.addAll(iapKeys)
-            keys.addAll(subscriptionKeys)
-            mBillingService = AmazonBillingService(contextLocal, keys)
-        }
+        mBillingService = BillingService(contextLocal, iapKeys, subscriptionKeys)
     }
 
     /**
@@ -92,7 +77,7 @@ object IAPManager {
     }
 
     @JvmStatic
-    fun getBillingService(): BillingService {
+    fun getBillingService(): IBillingService {
         return mBillingService ?: let {
             throw RuntimeException("Call IAPManager.build to initialize billing service")
         }
